getModel corrected

+++ administrator/components/com_cookiemanager/src/View/Cookies/HtmlView.php
@@ -145,7 +145,7 @@ class HtmlView extends BaseHtmlView
 			}
 		}
 
-		if ($this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
+		if (!$this->isEmptyState && $this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
 		{
 			$toolbar->delete('cookies.delete')
 				->text('JTOOLBAR_EMPTY_TRASH')
